+++ RELEASE_NOTES.md
@@ -1,3 +1,12 @@
+#### 5.207.3 - 2019-05-23
+* COSMETICS: Improved error message for missing monikers - https://github.com/fsprojects/Paket/pull/3583
+
+#### 5.207.2 - 2019-05-21
+* BUGFIX: Nuke project.assets.json files after paket install - https://github.com/fsprojects/Paket/issues/3577
+
+#### 5.207.1 - 2019-05-21
+* BUGFIX: Keep casing of packages stable in paket.lock - https://github.com/fsprojects/Paket/issues/3340
+
 #### 5.207.0 - 2019-05-11
 * Simplifier removes unsupported frameworks - https://github.com/fsprojects/Paket/pull/3574